fix: 파일 정리 및 rename
