+++ src/components/login/LoginHeader.jsx
@@ -1,28 +1,7 @@
 import React from 'react';
-import styled from 'styled-components';
 import { useLocation } from 'react-router-dom';
 import LogoImage from '../../assets/images/trends-logo.svg';
-
-const HeaderWrapper = styled.header`
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  position: relative;
-  width: 100%;
-  padding-top: 25px;
-  padding-left: 25px;
-`;
-
-const Logo = styled.img`
-  height: 7vh;
-`;
-
-const Divider = styled.div`
-  position: absolute;
-  width: 100%;
-  top: 14vh; /* 로고 아래로 더 내려감 */
-  border-top: 1px solid #cecece;
-`;
+import * as S from './LoginHeader.styles';
 
 const Header = () => {
   const location = useLocation();
@@ -34,10 +13,10 @@ const Header = () => {
 
   return (
     <>
-      <HeaderWrapper>
-        <Logo src={LogoImage} alt="유행 로고" />
-      </HeaderWrapper>
-      {showDivider && <Divider />}
+      <S.HeaderWrapper>
+        <S.Logo src={LogoImage} alt="유행 로고" />
+      </S.HeaderWrapper>
+      {showDivider && <S.Divider />}
     </>
   );
 };
